+++ test/barcode_scanner_plugin_method_channel_test.dart
@@ -7,21 +7,6 @@ void main() {
 
   MethodChannelBarcodeScannerPlugin platform =
       MethodChannelBarcodeScannerPlugin();
-  const MethodChannel channel = MethodChannel('barcode_scanner_plugin');
 
-  setUp(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, (MethodCall methodCall) async {
-      return '42';
-    });
-  });
-
-  tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
-  });
+  // Test có thể mở rộng ở đây để kiểm tra startCameraScan và barcodeStream
 }